Types update from int to Long on Entities' ids

+++ db-scripts/dbSchema.sql
@@ -18,7 +18,7 @@ USE `myexaminer` ;
 DROP TABLE IF EXISTS `myexaminer`.`account` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`account` (
-  `account_id` INT NOT NULL AUTO_INCREMENT,
+  `account_id` BIGINT NOT NULL AUTO_INCREMENT,
   `email` VARCHAR(100) NOT NULL,
   `password` VARCHAR(100) NOT NULL,
   `is_verified` BOOLEAN NOT NULL,
@@ -37,7 +37,7 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`role` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`role` (
-  `role_id` INT NOT NULL AUTO_INCREMENT,
+  `role_id` BIGINT NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(100) NOT NULL,
   PRIMARY KEY (`role_id`),
   UNIQUE INDEX `role_id_UNIQUE` (`role_id` ASC) VISIBLE,
@@ -51,7 +51,7 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`lecturer` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`lecturer` (
-  `fk_account_id` INT NOT NULL,
+  `fk_account_id` BIGINT NOT NULL,
   `first_name` VARCHAR(45) NOT NULL,
   `last_name` VARCHAR(45) NOT NULL,
   `home_page` VARCHAR(100) NOT NULL,
@@ -72,7 +72,7 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`student` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`student` (
-  `fk_account_id` INT NOT NULL,
+  `fk_account_id` BIGINT NOT NULL,
   `first_name` VARCHAR(45) NOT NULL,
   `last_name` VARCHAR(45) NOT NULL,
   `student_index` VARCHAR(6) NOT NULL,
@@ -94,11 +94,11 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`teaching_group` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`teaching_group` (
-  teaching_group_id INT NOT NULL AUTO_INCREMENT,
+  teaching_group_id BIGINT NOT NULL AUTO_INCREMENT,
   teaching_group_name VARCHAR(100) NOT NULL,
   access_code VARCHAR(45) NOT NULL,
   starting_date DATETIME NOT NULL,
-  fk_lecturer_account_id INT NOT NULL,
+  fk_lecturer_account_id BIGINT NOT NULL,
   PRIMARY KEY (`teaching_group_id`),
   UNIQUE (`teaching_group_id`),
   UNIQUE (`teaching_group_name`),
@@ -112,8 +112,8 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`student_teaching_group` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`student_teaching_group` (
-  `fk_teaching_group_id` INT NOT NULL,
-  `fk_student_account_id` INT NOT NULL,
+  `fk_teaching_group_id` BIGINT NOT NULL,
+  `fk_student_account_id` BIGINT NOT NULL,
   PRIMARY KEY (`fk_teaching_group_id`, `fk_student_account_id`),
   KEY `fk_teaching_group_id` (`fk_teaching_group_id`),
   CONSTRAINT `student_teaching_group_ibfk_2`
@@ -130,8 +130,8 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`account_role` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`account_role` (
-  `fk_account_id` INT NOT NULL,
-  `fk_role_id` INT NOT NULL,
+  `fk_account_id` BIGINT NOT NULL,
+  `fk_role_id` BIGINT NOT NULL,
   PRIMARY KEY (`fk_account_id`, `fk_role_id`),
   KEY `fk_role_id` (`fk_role_id`),
   CONSTRAINT `account_roles_ibfk_2`
@@ -147,13 +147,13 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`exam` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`exam` (
-  exam_id INT NOT NULL AUTO_INCREMENT,
+  exam_id BIGINT NOT NULL AUTO_INCREMENT,
   exam_name VARCHAR(50) NOT NULL,
   description VARCHAR(500),
   available_date DATETIME,
-  duration INT NOT NULL,
+  duration BIGINT NOT NULL,
   exam_status varchar(50),
-  fk_teaching_group_id INT,
+  fk_teaching_group_id BIGINT,
   PRIMARY KEY (`exam_id`),
   UNIQUE (`exam_id`),
   FOREIGN KEY (`fk_teaching_group_id`) REFERENCES `myexaminer`.`teaching_group`(`teaching_group_id`)
@@ -166,9 +166,9 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`exercise` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`exercise` (
-  exercise_id INT NOT NULL AUTO_INCREMENT,
+  exercise_id BIGINT NOT NULL AUTO_INCREMENT,
   exercise_body JSON NOT NULL,
-  fk_exam_id INT NOT NULL,
+  fk_exam_id BIGINT NOT NULL,
   PRIMARY KEY (`exercise_id`),
   UNIQUE (`exercise_id`),
   FOREIGN KEY (`fk_exam_id`) REFERENCES `myexaminer`.`exam`(`exam_id`)
@@ -178,9 +178,9 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`individual_exam` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`individual_exam` (
-  individual_exam_id INT NOT NULL AUTO_INCREMENT,
-  fk_exam_id INT NOT NULL,
-  fk_student_id INT NOT NULL,
+  individual_exam_id BIGINT NOT NULL AUTO_INCREMENT,
+  fk_exam_id BIGINT NOT NULL,
+  fk_student_id BIGINT NOT NULL,
   `is_checked` BOOLEAN NOT NULL,
   PRIMARY KEY (`individual_exam_id`),
   FOREIGN KEY (`fk_exam_id`) REFERENCES `myexaminer`.`exam`(`exam_id`),
@@ -191,9 +191,9 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`archive_exercise` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`archive_exercise` (
-  archive_exercise_id INT NOT NULL AUTO_INCREMENT,
-  fk_exercise_id INT NOT NULL,
-  fk_individual_exam_id INT NOT NULL,
+  archive_exercise_id BIGINT NOT NULL AUTO_INCREMENT,
+  fk_exercise_id BIGINT NOT NULL,
+  fk_individual_exam_id BIGINT NOT NULL,
   gained_points INT,
   answer JSON,
   lecturer_comment VARCHAR(500),
@@ -209,9 +209,9 @@ ENGINE = InnoDB;
 DROP TABLE IF EXISTS `myexaminer`.`notebook` ;
 
 CREATE TABLE IF NOT EXISTS `myexaminer`.`notebook` (
-  `notebook_id` INT NOT NULL AUTO_INCREMENT,
+  `notebook_id` BIGINT NOT NULL AUTO_INCREMENT,
   `notebook_body` LONGTEXT NULL,
-  `fk_account_id` INT NOT NULL,
+  `fk_account_id` BIGINT NOT NULL,
   PRIMARY KEY (`notebook_id`),
   UNIQUE INDEX `fk_account_id_UNIQUE` (`fk_account_id` ASC) VISIBLE,
   CONSTRAINT `fk_Student_Account2`
