+++ myexaminer/src/main/java/com/myexaminer/repository/LecturerRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LecturerRepository extends JpaRepository<Lecturer, Integer> {
-    long deleteByIdLecturer(int idLecturer);
-    Optional<Lecturer> findByIdLecturer(int idLecturer);
+public interface LecturerRepository extends JpaRepository<Lecturer, Long> {
+    long deleteByIdLecturer(Long idLecturer);
+    Optional<Lecturer> findByIdLecturer(Long idLecturer);
     Optional<Lecturer> findByAccount_Email(String email);
-    Optional<Lecturer> findByAccountId(int accountId);
+    Optional<Lecturer> findByAccountId(Long accountId);
 }
